requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beanie==1.25.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
lazy-model==0.2.0
motor==3.3.2
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
toml==0.10.2
typing_extensions==4.10.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
